Refactor gas parameter handling by introducing EthereumGasHelper
