+++ data/src/main/kotlin/com/vultisig/wallet/data/chains/helpers/ERC20Helper.kt
@@ -3,7 +3,6 @@ package com.vultisig.wallet.data.chains.helpers
 import com.google.protobuf.ByteString
 import com.vultisig.wallet.data.common.toKeccak256
 import com.vultisig.wallet.data.crypto.checkError
-import com.vultisig.wallet.data.models.Chain
 import com.vultisig.wallet.data.models.SignedTransactionResult
 import com.vultisig.wallet.data.models.payload.BlockChainSpecific
 import com.vultisig.wallet.data.models.payload.KeysignPayload
@@ -38,46 +37,16 @@ class ERC20Helper(
                     .build()
             )
 
-        return setGasParameters(
+        return EthereumGasHelper.setGasParameters(
             ethSpecific.gasLimit,
             ethSpecific.maxFeePerGasWei,
             input,
-            keysignPayload
+            keysignPayload,
+            nonceIncrement = BigInteger.ZERO,
+            coinType = coinType
         ).build().toByteArray()
     }
 
-    private fun requireEthereumSpec(spec: BlockChainSpecific): BlockChainSpecific.Ethereum =
-        spec as? BlockChainSpecific.Ethereum
-            ?: error("BlockChainSpecific is not Ethereum for EVM chain")
-
-    fun setGasParameters(
-        gas: BigInteger,
-        gasPrice: BigInteger,
-        signingInput: Ethereum.SigningInput.Builder,
-        keysignPayload: KeysignPayload,
-        nonceIncrement: BigInteger = BigInteger.ZERO,
-    ): Ethereum.SigningInput.Builder {
-        val ethSpecifc = requireEthereumSpec(keysignPayload.blockChainSpecific)
-        val signingInputBuilder = signingInput.apply {
-            chainId = ByteString.copyFrom(BigInteger(coinType.chainId()).toByteArray())
-            nonce = ByteString.copyFrom((ethSpecifc.nonce + nonceIncrement).toByteArray())
-        }
-        if(keysignPayload.coin.chain == Chain.BscChain) {
-            signingInputBuilder.apply  {
-                txMode = Ethereum.TransactionMode.Legacy
-                setGasPrice(ByteString.copyFrom(gasPrice.toByteArray()))
-                gasLimit =ByteString.copyFrom(gas.toByteArray())
-            }
-        } else {
-            signingInputBuilder.apply {
-                txMode = Ethereum.TransactionMode.Enveloped
-                gasLimit = ByteString.copyFrom(ethSpecifc.gasLimit.toByteArray())
-                maxFeePerGas = ByteString.copyFrom(ethSpecifc.maxFeePerGasWei.toByteArray())
-                maxInclusionFeePerGas = ByteString.copyFrom(ethSpecifc.priorityFeeWei.toByteArray())
-            }
-        }
-        return signingInputBuilder
-    }
     fun getPreSignedImageHash(keysignPayload: KeysignPayload): List<String> {
         val result = getPreSignedInputData(keysignPayload)
         val hashes = TransactionCompiler.preImageHashes(coinType, result)
