+++ data/src/main/kotlin/com/vultisig/wallet/data/wallet/OneInchSwap.kt
@@ -2,6 +2,7 @@ package com.vultisig.wallet.data.wallet
 
 import com.google.protobuf.ByteString
 import com.vultisig.wallet.data.api.models.OneInchSwapQuoteJson
+import com.vultisig.wallet.data.chains.helpers.EthereumGasHelper
 import com.vultisig.wallet.data.chains.helpers.EvmHelper
 import com.vultisig.wallet.data.common.toHexBytesInByteString
 import com.vultisig.wallet.data.models.OneInchSwapPayloadJson
@@ -65,16 +66,13 @@ class OneInchSwap(
         val gasPrice = quote.tx.gasPrice.toBigInteger()
         val gas = (quote.tx.gas.takeIf { it != 0L }
             ?: EvmHelper.DEFAULT_ETH_SWAP_GAS_UNIT).toBigInteger()
-        return EvmHelper(
-            keysignPayload.coin.coinType,
-            vaultHexPublicKey,
-            vaultHexChainCode,
-        ).setGasParameters(
+        return EthereumGasHelper.setGasParameters(
             gas = gas,
             gasPrice = gasPrice,
             signingInput = input,
             keysignPayload = keysignPayload,
             nonceIncrement = nonceIncrement,
+            coinType = keysignPayload.coin.coinType
         ).build().toByteArray()
     }
